+++ data/src/main/java/com/dida/data/model/response/login/GetCommonWalletResponse.kt
@@ -1,7 +1,7 @@
-package com.dida.data.model.response
+package com.dida.data.model.response.login
 
 import com.google.gson.annotations.SerializedName
 
-data class GetWalletExistsResponse(
+data class GetCommonWalletResponse(
     @SerializedName("existed") val existed: Boolean
 )
